Complete pagination, filters, sort in getAllProducts in catalog service.

+++ Services/Catalog/Catalog.Core/SpecsParams/Pagination.cs
@@ -1,10 +1,10 @@
 ﻿namespace Catalog.Core.SpecsParams;
 
-public class Pagination<T>(int pageIndex, int pageSize, int count, IReadOnlyList<T> dataList)
+public class Pagination<T>(int pageIndex, int pageSize, long count, IReadOnlyList<T> dataList)
     where T : class 
 {
     public int PageIndex { get; set; } = pageIndex;
     public int PageSize { get; set; } = pageSize;
-    public int Count { get; set; } = count;
+    public long Count { get; set; } = count;
     public IReadOnlyList<T> DataList { get; set; } = dataList;
 }
